refactor: Cleanup remainings
